+++ fluxo-kmp-conf/src/main/kotlin/SetupKotlin.kt
@@ -4,14 +4,20 @@ import fluxo.conf.impl.kotlin.hasKsp
 import org.gradle.api.Project
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmProjectExtension
 
+// TODO: Support JVM application with `application` plugin
+
 public fun Project.setupKotlin(
     config: (FluxoConfigurationExtension.() -> Unit)? = null,
     setupKsp: Boolean? = if (hasKsp) true else null,
     optIns: List<String>? = null,
     body: (KotlinJvmProjectExtension.() -> Unit)? = null,
 ): Unit = fluxoConfiguration {
-    if (setupKsp != null) this.setupKsp = setupKsp
-    if (!optIns.isNullOrEmpty()) this.optIns += optIns
+    if (setupKsp != null) {
+        this.setupKsp = setupKsp
+    }
+    if (!optIns.isNullOrEmpty()) {
+        this.optIns += optIns
+    }
     config?.invoke(this)
 
     asJvm {