+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/impl/kotlin/SetupJvmCompatibility.kt
@@ -109,6 +109,8 @@ internal const val JRE_1_9 = 9
 
 internal const val JRE_11 = 11
 
+internal const val JRE_21 = 21
+
 // https://www.oracle.com/java/technologies/downloads/
 private const val LTS_JDK_VERSION = 17